python/netsage_tools/tools/intelligence.py
import socket
import logging

def geoip_lookup(ip):
    """Perform a GeoIP lookup (Mock implementation for now)."""
    # In a real implementation, we'd use geoip2 or a REST API
    return {
        "status": "success",
        "ip": ip,
        "country": "United States",
        "city": "Mountain View",
        "latitude": 37.386,
        "longitude": -122.083,
        "isp": "Google LLC"
    }

def asn_lookup(ip):
    """Perform an ASN lookup (Mock implementation)."""
    return {
        "status": "success",
        "ip": ip,
        "asn": "AS15169",
        "org": "Google LLC"
    }

def whois_lookup(domain):
    """Perform a WHOIS lookup."""
    try:
        import whois
        w = whois.whois(domain)
        return {
            "status": "success",
            "domain": domain,
            "registrar": w.registrar,
            "creation_date": str(w.creation_date),
            "expiration_date": str(w.expiration_date)
        }
    except Exception as e:
        return {
            "status": "error",
            "message": f"WHOIS failed: {str(e)}"
        }
